Create extension function
